requirements.txt
torch
torchvision
pytorch-lightning
diffusers
mlflow
torch-fidelity
numpy
pillow
scikit-learn
fastapi
uvicorn
streamlit
requests

